requirements.txt
Django==2.0.2
djangorestframework==3.7.7
numpy==1.14.1
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.0